got travis.yml search working and saving
search_code is needed in order to get the filename:query search working
as that doesn't work on search_repos... for some reason. The docs are a
little vague but I what they mean is that repo searching can only be
done on name and description which kind of makes sense.

Note: encoding bug needs to sorted along with getting the config into
the csv file as well. Luckily that will be base64 encoded as well so
should be pretty simple to do