fixed bug and added in circleci back in
bug in python library encountered when reading files that should be
directories. There is a pending fix which converts get_contents() and
get_dir_conents() (or something) to just get_contents().

We have been using get_contents() already so the code has been
refactored to handle that. In doing so has made PATHS a lot simpler
which is really nice!