bugs fixed and working comment stats and travis parsing
bugs:
- double quotes for yaml configurations wasn't being handled properly
for comments (leading to a false postive in results)
	- unit tests added to make sure this doesn't happen again
- jenkins configuration not being picked up and unit tests added to
scraper.py

data_parser.py now spits out another file with the contents of the
comments in there. This should allow for commments about TODO comments
to be stored.

travis parsing works but need to have a more focused approach in nowing
what we are going to do.