Fixed logging defaults, tidied up assumptions in search
Before hand we were assuming that a piece of yaml configuration was:
- config.yaml
but as we are now searching for that single file instead of using the
search api. We need to search for all possible files of that type e.g.
.config.yaml, .config.yml, config.yaml, config.yml.

Next commit should look at the implications that has on rate limiting.