got jenkins style comments working
data_parser has been changed to run temp which takes a parameter to
specifiy if it is yaml. Then picks the comment function to use dependant
on that. This used to be done by the processes method.

Processing comments now works for the single comment style of "#" and
the other style of "//" and "/*".