Change of tilte on the report and rewriting intro (added some ref)
TIL: that having comma's in the author field in a .bib file which is the file that is used to keep track of the references. It is actuallly for the ordering of speakers and so "and" should be used instead.

See: https://tex.stackexchange.com/questions/343163/error-in-bib-entry-too-many-commas

On another note this seems to be better feel for the paper with this
title and it might actually come in on time..... It's going to be
interesting comparing results with the other Usage paper though...